Refactoring for Admin folders
