+++ backend/src/main/java/com/example/Caramelca/services/Admin/AdminUserService.java
@@ -1,4 +1,4 @@
-package com.example.Caramelca.services;
+package com.example.Caramelca.services.Admin;
 
 import com.example.Caramelca.models.Role;
 import com.example.Caramelca.models.User;
